move user utility files
